DataGrid - Reorganize keyboard navigation tests (IE, Edge farm testing issues)
