+++ testing/tests/DevExpress.ui.widgets.dataGrid/keyboardNavigation.realControllers.tests.js
@@ -1,18 +1,28 @@
+QUnit.testStart(function() {
+    let markup = `
+        <div>
+            <div id="container" class="dx-datagrid"></div>
+        </div>`;
+
+    $('#qunit-fixture').html(markup);
+});
+
 import 'common.css!';
 import 'generic_light.css!';
 
+import $ from 'jquery';
+
 import 'ui/data_grid/ui.data_grid';
 
-import $ from 'jquery';
 import browser from 'core/utils/browser';
 import commonUtils from 'core/utils/common';
 import pointerEvents from 'events/pointer';
-import { setupDataGridModules } from '../../../helpers/dataGridMocks.js';
+import { setupDataGridModules } from '../../helpers/dataGridMocks.js';
 import {
     CLICK_EVENT,
     triggerKeyDown,
     focusCell,
-    dataGridWrapper } from '../../../helpers/grid/keyboardNavigationHelper.js';
+    dataGridWrapper } from '../../helpers/grid/keyboardNavigationHelper.js';
 
 QUnit.module('Real DataController and ColumnsController', {
     setupModule: function() {
