--- testing/tests/DevExpress.ui.widgets.dataGrid/keyboardNavigation.tests.js
@@ -1,16 +1,0 @@
-QUnit.testStart(function() {
-    let markup = `
-        <div>
-            <div id="container" class="dx-datagrid"></div>
-        </div>`;
-
-    $('#qunit-fixture').html(markup);
-});
-
-import $ from 'jquery';
-import './keyboardNavigationParts/keyboardController.tests.js';
-import './keyboardNavigationParts/keyboardKeys.tests.js';
-import './keyboardNavigationParts/rowsView.tests.js';
-import './keyboardNavigationParts/realControllers.tests.js';
-import './keyboardNavigationParts/customization.tests.js';
-import './keyboardNavigationParts/accessibility.tests.js';
